requirements.txt
numpy==1.24.1
pillow==9.4.0
pandas==1.5.2
matplotlib==3.6.2
opencv-contrib-python==4.7.0.68
mediapipe==0.9.0.1
pyautogui==0.9.53
scikit-learn==1.2.0
seaborn==0.12.2